Move profile image to profile command
